requirements.txt
# Core ML/DL
torch==2.1.0
transformers==4.36.2
datasets==2.16.1
accelerate==0.25.0
peft==0.7.1
bitsandbytes==0.41.3

# Training utilities
wandb==0.16.2
trl==0.7.10
scipy==1.11.4

# Evaluation
evaluate==0.4.1
rouge-score==0.1.2
nltk==3.8.1
bert-score==0.3.13

# Data processing
pandas==2.1.4
numpy==1.26.2
tqdm==4.66.1
scikit-learn==1.3.2

# Visualization
matplotlib==3.8.2
seaborn==0.13.0

# Utilities
pyyaml==6.0.1
python-dotenv==1.0.0
sentencepiece==0.1.99

# Development
jupyter==1.0.0
ipywidgets==8.1.1
black==23.12.1
pytest==7.4.3
